Fixed a Parse bug
A year a go a bug was discovered where when you get the user object,
you will get the user ID, but from time to time you won’t get other
fields. No error will be thrown etc. TO fix the issue, once you get the
user object, you have to use fetch to force the SDK to get it again (so
magic is going on there for sure). More about the issue can be found
here:
https://www.parse.com/questions/get-null-from-pfuser-currentuser-objectf
orkeyusername-unless-calling-fetch-in-appdelegatem